Refactor: 컨트롤러 리팩토링
* 게시글 생성 메서드의 호출 URL을 변경하고 Account 엔티티를 찾는 로직을 서비스 단으로 이동시킴에 따라 삭제함

* 기존의 ArticleController2를 ArticleController로 통합시킴
